feat: simplify authorization logic in request classes and add new policies for project and organization management

+++ backend_app/app/Http/Controllers/Organization/OrganizationController.php
@@ -13,12 +13,15 @@ use App\Models\Organization;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
+use Illuminate\Support\Facades\Gate;
 use Throwable;
 
 class OrganizationController extends Controller
 {
     public function index(Request $request): JsonResource
     {
+        Gate::authorize('viewAny', Organization::class);
+
         return new OrganizationCollection(
             $request->user()
                 ->ownedOrganizations()
@@ -31,25 +34,29 @@ class OrganizationController extends Controller
 
     public function store(StoreOrganizationRequest $request): JsonResource
     {
+        Gate::authorize('create', Organization::class);
+
         $organization = $request->user()
             ->ownedOrganizations()
-            ->with('owner')
             ->create($request->validated());
 
-        return new OrganizationResource($organization);
+        return new OrganizationResource($organization->load('owner'));
     }
 
     public function show(Organization $organization): JsonResource
     {
-        // todo add policy check
-        $organization->load('owner')->loadCount(['workspaces', 'users']);
+        Gate::authorize('view', $organization);
 
-        return new OrganizationResource($organization);
+        return new OrganizationResource($organization->load('owner')
+            ->loadCount(['workspaces', 'users'])
+        );
     }
 
     /** @throws Throwable */
     public function update(UpdateOrganizationRequest $request, Organization $organization): JsonResource
     {
+        Gate::authorize('update', $organization);
+
         $organization->updateOrFail($request->validated());
 
         return new OrganizationResource($organization->with('owner'));
@@ -58,6 +65,8 @@ class OrganizationController extends Controller
     /** @throws Throwable */
     public function destroy(Organization $organization): JsonResponse
     {
+        Gate::authorize('delete', $organization);
+
         $organization->deleteOrFail();
 
         return response()->json(status: 204);
